refactor(cast): inline block implementation
